% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_load_data.R
\name{clean_load_data}
\alias{clean_load_data}
\title{Clip and save waterbird data}
\usage{
clean_load_data(
  waterbird_data,
  basin,
  assets,
  valleys,
  nsbasin,
  remove_adhoc = TRUE
)
}
\arguments{
\item{waterbird_data}{dataframe with waterbird data}

\item{basin}{sf object of basin}

\item{assets}{sf object of wetland assets}

\item{valleys}{sf object of valleys}
}
\value{
a dataframe with waterbird data clipped to the basin, and joined with the assets, valleys, and nsbasin data
}
\description{
Clip and save waterbird data
}
