% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_model_output.R
\name{clean_model_output}
\alias{clean_model_output}
\title{Clean output from run_mk_models()}
\usage{
clean_model_output(output, decade_analysis, grouping_cols, by_condition = NULL)
}
\arguments{
\item{output}{output from run_mk_models()}

\item{decade_analysis}{logical indicating if decadal analysis was performed}

\item{grouping_cols}{character string indicating the grouping column used in the analysis (if any)}

\item{by_condition}{character string indicating the condition column used in the analysis (if any)}
}
\value{
A cleaned data frame with model results
}
\description{
Clean output from run_mk_models()
}
\examples{
\dontrun{
cleaned <- clean_model_output(
   output,
   decade_analysis = decade_analysis,
   grouping_cols = grouping_cols,
   by_condition = by_condition
 )
}
}
