% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_clean}
\alias{data_clean}
\title{Cleaned waterbird data produced by \code{clean_clip_raw_data()}}
\format{
\subsection{\code{data_clean}}{

A tibble with 53 columns
}
}
\usage{
data_clean
}
\description{
Cleaned waterbird data produced by \code{clean_clip_raw_data()}
}
\keyword{datasets}
