% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mk_model.R
\name{fit_mk_model}
\alias{fit_mk_model}
\title{Fit Mann-Kendall trend model.}
\usage{
fit_mk_model(data, decade_analysis)
}
\arguments{
\item{data}{Dataframe containing the data to fit the model on. Usually a nested dataframe.}

\item{decade_analysis}{Logical indicating if the analysis is decadal.}
}
\value{
A list containing the following elements:
\itemize{
\item model: The fitted Mann-Kendall model object.
\item tau: The Kendall's tau statistic from the model.
\item sl: The p-value associated with the tau statistic.
\item log_slope: The Theil-Sen slope estimate on the log-transformed data.
\item percent_slope: The percent slope, calculated from the log_slope.
\item intercept: The intercept of the Theil-Sen line on the log-transformed data.
\item warnings: A character vector of any warnings captured during model fitting.
\item has_warnings: A logical indicating if any warnings were captured.
}
}
\description{
For detection of trend in a time series, and a Theil-Sen slope estimate.
}
\examples{
\dontrun{

mk_output = purrr::map(
       data,
       fit_mk_with_warnings,
       decade_analysis = decade_analysis
     )
}
}
