% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsbasin}
\alias{nsbasin}
\title{Polygon of the MDB, split into North and South}
\format{
\subsection{\code{nsbasin}}{

A sf polygon
}
}
\usage{
nsbasin
}
\description{
Polygon of the MDB, split into North and South
}
\keyword{datasets}
