% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waterbird_all}
\alias{waterbird_all}
\title{Raw waterbird data from the EAWS and other surveys}
\format{
\subsection{\code{waterbird_all}}{

A tibble with 43 columns
}
}
\usage{
waterbird_all
}
\description{
Raw waterbird data from the EAWS and other surveys
}
\keyword{datasets}
