% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_clip_raw_data.R
\name{clean_clip_raw_data}
\alias{clean_clip_raw_data}
\title{Clip and clean raw waterbird data}
\usage{
clean_clip_raw_data(
  waterbird_data,
  basin,
  assets,
  valleys,
  nsbasin,
  remove_adhoc = TRUE
)
}
\arguments{
\item{waterbird_data}{dataframe with waterbird data. This should be the raw, unclipped waterbird data}

\item{basin}{sf object of Basin}

\item{assets}{sf object of wetland assets}

\item{valleys}{sf object of valleys (usually SRA valleys)}

\item{nsbasin}{sf object of basin divisions (north and south)}

\item{remove_adhoc}{Logical. Whether to remove adhoc surveys. Default TRUE.}
}
\value{
a dataframe with waterbird data clipped to the Basin, and joined with the assets, valleys, and nsbasin data
}
\description{
This function takes the raw waterbird data, clips it to the Basin, and joins it with the assets, valleys, and nsbasin data, so that each point is assigned to the wetland, valley, and basin it belongs to.
It also has an option to remove adhoc surveys, which should only be used for basin scale analysis of EAWS data.
}
