% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_run_models.R
\name{clean_data_run_models}
\alias{clean_data_run_models}
\title{Main function which cleans data, runs models, and cleans output}
\usage{
clean_data_run_models(
  dataset = data_clean,
  programs,
  metric,
  wetlands,
  valleys,
  basinDiv,
  grouping_cols,
  decade_analysis = FALSE,
  by_condition = NULL
)
}
\arguments{
\item{dataset}{Dataset for analysis. Should be output from \code{clean_clip_raw_data()} (or default \code{data_clean}).}

\item{programs}{Which survey programs to include}

\item{metric}{Metric to analyse. One of "abundance", "richness", "nests", "broods", "nest_richness", or "pct_filled"}

\item{wetlands}{Which wetlands to include (default all)}

\item{valleys}{Which valleys to include (default all)}

\item{basinDiv}{Character string indicating north or south basin}

\item{grouping_cols}{Column to group by (e.g. "ValleyName", "Wetland", "spp_code"). Default NULL}

\item{decade_analysis}{Logical. If TRUE, runs models separately for each decade}

\item{by_condition}{character string indicating the condition column used in the analysis (if any). This should be one of "wetland_condition", "valley_condition", "basin_condition", or "division_condition". Default NULL. If specified, this will group by the condition of the wetland/valley/basin/division that each survey point belongs to. However, this does rely on the SWC of those units being joined to the dataset. It will rarely be used.}
}
\value{
A list containing:
\enumerate{
\item A cleaned data frame with model results
\item A ggplot object with the post-fitting line plot
\item A summary data frame with summary statistics and model results by group (if
}
}
\description{
Main function which cleans data, runs models, and cleans output
}
