% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eawsAnalysis-package.R
\docType{package}
\name{eawsAnalysis-package}
\alias{eawsAnalysis}
\alias{eawsAnalysis-package}
\title{eawsAnalysis: This package is used to analyse waterbird data from the Eastern Australian Waterbird Survey, and other associated waterbird surveys}
\description{
This package provides an interface to the data from the Eastern Australian Waterbird Survey, and other associated waterbird surveys, and functions to analyse these data, including maps, summary tables, and trend analyses.
}
\author{
\strong{Maintainer}: Will Wright \email{will.wright@unsw.edu.au}

Authors:
\itemize{
  \item Will Wright \email{will.wright@unsw.edu.au}
}

}
\keyword{internal}
