% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bird_data.R
\name{get_bird_data}
\alias{get_bird_data}
\title{Extract cleaned waterbird data}
\usage{
get_bird_data(
  df = clipped_joined_dataset,
  programs,
  wetlands,
  valleys,
  basinDiv,
  grouping_cols,
  metric = "abundance",
  by_condition = NULL
)
}
\arguments{
\item{df}{cleaned and joined dataframe}

\item{programs}{Which survey programs to include}

\item{wetlands}{which wetlands to include}

\item{valleys}{which valleys to include}

\item{grouping_cols}{Column to group by (e.g. "ValleyName", "Wetland", "spp_code"). Default NULL}

\item{metric}{Metric to analyse. One of "abundance", "richness", "nests", "broods", "nest_richness", or "pct_filled"}
}
\value{
a dataframe with the specified metric calculated for each group and year
}
\description{
Extract cleaned waterbird data
}
\examples{
--- IGNORE ---
}
