% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_fitting_line_plot.R
\name{post_fitting_line_plot}
\alias{post_fitting_line_plot}
\title{Generate post-fitting line plot}
\usage{
post_fitting_line_plot(data, decade_analysis, grouping_cols, metric)
}
\arguments{
\item{data}{Cleaned data from \code{clean_model_output()}}

\item{decade_analysis}{Logical indicating if decadal analysis was performed}

\item{grouping_cols}{Character string indicating grouping column used (if any)}

\item{metric}{Metric name for y-axis label}
}
\description{
Generate post-fitting line plot
}
