% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mk_models.R
\name{run_mk_models}
\alias{run_mk_models}
\title{Run trend models models on cleaned dataset from \code{get_bird_data()}, using \code{fit_mk_model()} for Mann-Kendall trend detection and Theil-Sen slope estimation.}
\usage{
run_mk_models(
  cleaned_dataset,
  grouping_cols = NULL,
  decade_analysis = FALSE,
  by_condition = NULL
)
}
\arguments{
\item{cleaned_dataset}{A cleaned dataset obtained from the \code{clean_data} function.}

\item{grouping_cols}{An optional character string specifying a column name in the dataset to group by before modeling. E.g. "spp_code" or "ValleyName".}

\item{decade_analysis}{A logical indicating whether to perform decade-wise analysis. If TRUE, models are fitted separately for each decade.}

\item{by_condition}{by_condition whether to group by surface water condition. This should be one of "wetland_condition", "valley_condition", "basin_condition", or "division_condition". Default NULL. If specified, this will group by the condition of the wetland/valley/basin/division that each survey point belongs to. However, this does rely on the SWC of those units being joined to the dataset. It will rarely be used.}
}
\value{
A nested tibble containing the original data, model results, and warnings (if grouping_cols is provided) or overall.
}
\description{
Run Mann-Kendall trend analyses on a dataset.

This function relies on the \code{Kendall} package (run with the helper function \code{fit_mk_model()})
to perform the Mann-Kendall test. This function also provides a Theil-Sen median slope estimate.
}
\examples{
\dontrun{
output <- run_mk_models(
   cleaned_data,
   decade_analysis = decade_analysis,
   family_for_glm = family_for_glm,
   grouping_cols = grouping_cols,
   by_condition = by_condition
)
}


}
